Extract rollover tests into their own test class (#69295)
